Don't inject ActiveRecord::Base for sort/serialization
ActiveRecord 5 adds a check for descendants for single table inheritance
in a new place, which makes fixture_builder fail on any STI models that
have this ActiveRecord::Base injected into it. The class that is
injected has no descendants, and relies on only the table name.

New check:
https://github.com/rails/rails/blob/master/activerecord/lib/active_record/inheritance.rb#L203-L205

Instead, have the existing procs take in just the records pulled out of
the database tables and deal with them that way. This still allows for
our serialization and sorting, just in a more rudimentary way.